+++ tests/sanity/tests/actions.spec.ts
@@ -13,27 +13,23 @@ test.describe('actions tests', () => {
   test('action-new-candidate', async ({ page }) => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
-    await page.click('text=Candidates')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
-    )
+    await page.click('text=Talents')
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
 
     await page.click('td:has-text("Frontend Engineer")')
 
     await page.press('body', 'Meta+k')
-    await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'Candidate')
-    expect(await page.locator('div.selectPopup :text("New Candidate")').count()).toBe(1)
-    await page.click('div.selectPopup :text("New Candidate")')
+    await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'Talent')
+    expect(await page.locator('div.selectPopup :text("New Talent")').count()).toBe(1)
+    await page.click('div.selectPopup :text("New Talent")')
     await page.click('button#card-close')
   })
 
   test('action-switch-vacancies', async ({ page }) => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
-    await page.click('text=Candidates')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
-    )
+    await page.click('text=Talents')
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
 
     await page.press('body', 'Meta+k')
     await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'go to')
@@ -47,10 +43,8 @@ test.describe('actions tests', () => {
   test('action-switch-applications', async ({ page }) => {
     await page.click('[id="app-recruit\\:string\\:RecruitApplication"]')
 
-    await page.click('text=Candidates')
-    await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
-    )
+    await page.click('text=Talents')
+    await expect(page).toHaveURL(`${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/talents`)
 
     await page.press('body', 'Meta+k')
     await page.fill('[placeholder="type\\ to\\ filter\\.\\.\\."]', 'go to')
@@ -58,7 +52,7 @@ test.describe('actions tests', () => {
     await page.click('div.selectPopup :text("Go To Applications")')
 
     await expect(page).toHaveURL(
-      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/applicants`
+      `${PlatformURI}/workbench%3Acomponent%3AWorkbenchApp/recruit%3Aapp%3ARecruit/candidates`
     )
   })
 })